// mlt.web/src/app/core/enums/DownloadStatus.ts
export enum DownloadStatus {
  Waiting = 'waiting',
  Downloading = 'downloading',
  Paused = 'paused',
  Finishing = 'finishing',
  Finished = 'finished',
  HashChecking = 'hash_checking',
  Seeding = 'seeding',
  FilehostingWaiting = 'filehosting_waiting',
  Extracting = 'extracting',
  Error = 'error'
}